fix(ui): back to bootstrap
Primeng is too large and complex for my needs. You need to prepare datastructures for ui and it's a bit anoying... Goal of this app is to force myself to use angular forms in advance way, validate, validate with api, show errors, autocomplete and async validators instead of messing around with shiny uikit. A few more words about bootstrap - it's minify version weitght around 250 kB - it's less then Bulma and SemanricCSS. It's simple and flexible enough to construct new components, and less css classes then Tailwind ( wich is great but... this lines of css classes ... ). So back to bootstrap again.